NumericalMethods

Documentation for NumericalMethods.

NumericalMethods

Functions

Index

Functions

NumericalMethods.ODE — Type

ODE(f, a, b, h, α, β, N)

Structure of the boundary conditions to differential equation.

Notes

Make sure the independent variable (e.g. time) is the first argument of f!

NumericalMethods.SVI — Type

SVI(f, a, b, n, tol)

Given f(x) such that x ∈ [a, b], find the root of a single-variable, equation in so many iterations, n within tolerance, tol.

Methods

bisection() Search for solution by halving the bounds wherein a and b initially yield opposite signs in function. falseposition(p0: float, p1: float) solution bounded by a and b wherein initial guesses p0 and p1 yield opposite signs in function. fixedpoint(p0: float) Root-finding method to find solution near initial guess. newtonraphson(p0: float) Root-finding method to find solution near initial guess. secantmethod(p0: float, p1: float) Initial guesses p0 and p1 must yield opposite signs in function. Solution is NOT bounded by a and b.

Notes

Convergence Rates:

`newton_raphson` > `secant_method` > `false_position` > `fixed_point` > `bisection`

NumericalMethods.bezier — Method

bezier(x, y, xguides, yguides)

An application of Hermitic polynomials to draw Bezier curves between points.

Notes

Each argument should be a one-to-one mapping of points, (xᵢ, yᵢ) and (xᵢ₊₁, yᵢ₊₁) and their respective guide points, (xᵢ⁺, yᵢ⁺) and (xᵢ₊₁⁻, yᵢ₊₁⁻).

NumericalMethods.bisection — Method

bisection(obj::SVI)

Root-finding method: f(x) = 0. Search for solution by halving the bounds wherein a and b initially yield opposite signs in function.

Notes

Relying on the Intermediate Value Theorem (IVT), this is a bracketed, root-finding method. This method is rather slow to converge but will always converge to a solution; therefore, is a good starter method.

NumericalMethods.clamped — Method

clamped(x, f, fp)

The bookend polynomials will have the same slope entering and exiting the interval as the derivative at the respective endpoint.

NumericalMethods.endpoint — Method

endpoint(x, f, h, point[; method=:three])

Find the derivative of a bookend point at either :begin or :end of dataset. Acceptable values for method include {:three, :five}.

NumericalMethods.false_position — Method

false_position(obj, p0, p1)

Attempt root-finding method with initial guesses, p0 and p1 in [a, b] solving the equation g(p) = p via f(p) - p = 0.

Use function with lowest slope!

Notes

Similar to, but slower to converge than, the secant_method() by including a test to ensure solution is root-bracketed.

See fixed_point() for theorem.

NumericalMethods.fixed_point — Method

fixed_point(obj, p0)

Attempt root-finding method with initial guess, p0 in [a, b] by solving the equation g(p) = p via f(p) - p = 0.

Use function with lowest slope!

Not root-bracketed.

Notes

Theorem:

Existence of a fixed-point: If g ∈ C[a,b] and g(x) ∈ C[a, b] for all x ∈ [a, b], then function, g has a fixed point, p ∈ [a, b].

Uniqueness of a fixed point: If g'(x) exists on [a, b] and a positive constant, k < 1 exist with {|g'(x)| ≤ k | x ∈ (a, b)}, then there is exactly one fixed-point, p ∈ [a, b].

Converges by mathcal{O}(text{linear}) if g'(p) ≠ 0, and mathcal{O}(text{quadratic}) if g'(p) = 0 and g''(p) < M, where M = g''(ξ) that is the error function.

NumericalMethods.integrate — Method

integrate(f, x      [; rule=:trapezoidal, tol=10^-3])
integrate(f, a, b, h[; rule=:trapezoidal, tol=10^-3])
integrate(f, a, b, n[; rule=:trapezoidal, tol=10^-3])

Find the definite integral by some numerical quadrature.

Notes

f may be a function or range. The domain may be defined with a vector, x or on the interval [a, b] either by number of sub-intervals, n or step-size, h. rule accepts {:trapezoidal (default), :midpoint, :simpson13, :simpson38, :simpsonN}. Dataset may contain unevenly spaces points.

References

https://en.wikipedia.org/wiki/Simpson%27s_rule

NumericalMethods.ivp — Method

ivp(obj::ODE[; tol=10^-3, method=:forward_euler])

Solve obj according to method ∈ {:forward_euler (default), :backward_euler, :improved_euler, :modified_euler, :runge_kutta}.

Notes

Each method has an equivalent convenience function. E.g. ivp(obj; method=:runge_kutta) ≡ runge_kutta(obj).

NumericalMethods.lagrange — Method

lagrange(x, f[; n=nothing])

Given a domain, x and range, f, construct the nth Lagrangian polynomial.

Notes

If n=nothing, then method will utilize entire dataset. Polynomials will quickly oscillate for larger datasets.

NumericalMethods.linearleastsquares — Method

linearleastsquares(x, f, n::Integer)

Construct a polynomial of degree, n while minimizing the least squares error.

Notes

Least squares error := $E = \sum_{i=1}^{m}[y_{i} - P_{n}(x_{i})]^{2}$

Constructed polynomial of the form: $P(x) = a_{n}x^{n} + a_{n - 1}x^{n - 1} + \dots + a_{1}x + a_{0}$

NumericalMethods.linearleastsquares — Method

linearleastsquares(x, f, type::Symbol)

Given a domain and range, yield the coefficients for an equation and the equation of the form $y = ax^{b}$.

NumericalMethods.max_iterations — Function

max_iterations(obj, method[, p0=0])

Find greatest integer for maximum iterations within tolerance.

Notes

Acceptable values for method ∈ {:bisection, :fixed_point, :newton_raphson, :secant_method, :false_position}. Initial guess, p0 for function solution is not needed for :bisection method.

NumericalMethods.maximumslope — Method

maximumslope(obj::SVI)

Find the greatest value for first derivative of function.

NumericalMethods.midpoint — Method

midpoint(x, f, h, point[; method=:three])

Find the derivative of some point within a dataset. Acceptable values for method include {:three, :five, :2nd}.

NumericalMethods.n1derivative — Method

n1derivative(x, f, j[; n=nothing])

The general (n + 1)-point formula to approximate f' at point j.

Notes

If n = nothing, then entire dataset used to construct nth Lagrange coefficient.

NumericalMethods.natural — Method

natural(x, f)

The bookend polynomials do not assume the slope entering and exiting the interval as the derivative at the respective endpoint.

NumericalMethods.newton_raphson — Method

newton_raphson(obj, p0)

Attempt root-finding method with initial guess, p0 in [a, b] by solving the equation g(p) = p via f(p) - p = 0.

Use function with lowest slope!

f'(x) ≠ 0

Notes

Not root-bracketed and has trouble with symmetric functions! Initial guess, p0 must be close to real solution; else, will converge to different root or oscillate (if symmetric). This method can be viewed as fixed-point iteration.

Technique based on first Taylor polynomial expansion of f about p{0} (that is p0) and evaluated at x = p. |p - p{0}| is assumed small; therefore, 2^{ ext{nd}}-order Taylor term, the error, is small.

Newton-Raphson has quickest convergence rate.

See fixed_point() for theorem.

NumericalMethods.newtondifference — Method

newtondifference(x, f, α[; dir::Symbol=:auto])

Given a domain, x and range, f, construct some polynomial by Newton's Divided Difference centered around α. :forward or :backward construction.

Notes

Direction will be chosen if not specified. Polynomials best made with even spacing in x; although, this is not completely necessary.

NumericalMethods.secant_method — Method

secant_method(obj, p0, p1)

Attempt root-finding method with initial guesses, p0 and p1 in [a, b] by solving the equation g(p) = p via f(p) - p = 0.

Use function with lowest slope!

Not root-bracketed.

Notes

Method is less computationally expensive than newton_raphson() but may converge at slower rate by circumventing need to calculate derivative.

See fixed_point() for theorem.

Index

NumericalMethods.ODE

NumericalMethods.SVI

NumericalMethods.bezier

NumericalMethods.bisection

NumericalMethods.clamped

NumericalMethods.endpoint

NumericalMethods.false_position

NumericalMethods.fixed_point

NumericalMethods.integrate

NumericalMethods.ivp

NumericalMethods.lagrange

NumericalMethods.linearleastsquares

NumericalMethods.linearleastsquares

NumericalMethods.max_iterations

NumericalMethods.maximumslope

NumericalMethods.midpoint

NumericalMethods.n1derivative

NumericalMethods.natural

NumericalMethods.newton_raphson

NumericalMethods.newtondifference

NumericalMethods.secant_method