Eigenvalue

Types

NumericalMethods.Eigenvalues.Eigenvalue — Type

Eigenvalue(A, N, tol)

Find the characteristic (eigen) values of matrix, A. Typically thought as roots of polynomial from determinant.

Functions

NumericalMethods.Eigenvalues.inverse_power_method — Method

inverse_power_method(EV, x0, q)

Approximate eigenvalue closest to target, q of matrix, EV.A given some non-zero, initial guess vector, x0. Supposed to converge faster than power_method [burdenNumericalAnalysis2016]_.

NumericalMethods.Eigenvalues.power_method — Method

power_method(EV, x0)

Approximate the dominant eigenvalue of matrix, EV.A given some non-zero, initial guess vector, x0.

NumericalMethods.Eigenvalues.qr_algorithm — Method

qr_algorithm(EV)

Systematically find all the eigenvalues of matrix, EV.A.

Matrix must be symmetric and tridiagonal!

This method is preferred over power_method and inverse_power_method by keeping round-off error to a minimum [burdenNumericalAnalysis2016]_. Refer to this example for an explanation and demonstration.

Index

NumericalMethods.Eigenvalues.Eigenvalue

NumericalMethods.Eigenvalues.inverse_power_method

NumericalMethods.Eigenvalues.power_method

NumericalMethods.Eigenvalues.qr_algorithm