Multi-Variable Iteration

Types

NumericalMethods.MultiVariableIterations.MultiVariableIteration — Type

MultiVariableIteration(A, x, b, N, tol)

Given the system of equations, $\mathbf{A}\vec{x} = \vec{b}$, attempt to solve $\vec{x} = \mathbf{A}^{-1}\vec{b}$ in so many iterations, N within tolerance, tol. Ideal for large, sparse systems.

Functions

NumericalMethods.MultiVariableIterations.condition_number — Method

condition_number(A)

Find the condition number of matrix, A.

Notes

Definition [burdenNumericalAnalysis2016]_: The condition number of the non-singular matrix, $\mathbf{A}$ relative to a norm, ||⋅|| is

\[ K(\mathbf{A}) = ||\mathbf{A}|| ⋅ ||\mathbf{A}^{-1}||\]

A matrix is well-conditioned if $K(\mathbf{A})$ is close to 1 and is ill-conditioned if significantly greater than 1.

NumericalMethods.MultiVariableIterations.diagonality — Method

diagonality(A)

Determines whether matrix, A is strictly, diagonally dominant.

NumericalMethods.MultiVariableIterations.find_omega — Function

find_omega(mvi[, omega=0.])

Find the optimum relaxation parameter, omega for mvi.A and mvi.x, if possible.

Notes

If 0 < omega < 2, then method will converge regardless of choice for mvi.x. If matrix, mvi.A is tridiagonal, then the spectral radius of Gauss-Seidel's T-matrix, $\mathbf{T}\_{g}$ is used to calculate $ω := 2 / (1 + \sqrt{(1 - `spectral_radius`(\mathbf{T}\_{g})})$ [burdenNumericalAnalysis2016]_.

NumericalMethods.MultiVariableIterations.gauss_seidel — Method

gauss_seidel(mvi)

Solve $\vec{x} = \mathbf{A}^{-1}\vec{b}$ via the Gauss-Seidel Method.

Notes

This method improves on jacobi() by using the most recently calculated entries in the approximation vector, x at the end of each iteration. The core algorithm by which method marches through iterations:

\[ \vec{x}^{(k)} = \bigl( (\mathbf{D} - \mathbf{L})^{-1} * \mathbf{U} \bigr) ⋅ \vec{x}^{(k - 1)} + \bigl( (\mathbf{D} - \mathbf{L})^{-1} \bigr) ⋅ \vec{b}\]

NumericalMethods.MultiVariableIterations.jacobi — Method

jacobi(mvi)

Solve $\vec{x} = \mathbf{A}^{-1}\vec{b}$ via the Jacobi Method to find $\vec{x}$.

Notes

The core algorithm by which method marches through iterations:

\[ \vec{x}^{(k)} = \bigl( \mathbf{D}^{-1} * (\mathbf{L} + \mathbf{U}) \bigr) ⋅ \vec{x}^{(k - 1)} + ( \mathbf{D}^{-1} ) ⋅ \vec{b}\]

NumericalMethods.MultiVariableIterations.positive_definite — Method

positive_definite(A)

Determines whether matrix, A is positive definite.

NumericalMethods.MultiVariableIterations.spectral_radius — Method

spectral_radius(A)

Finds the spectral radius of matrix, A.

Notes

$ρ(\mathbf{A}) = \max|λ|$, where λ is the set of eigvalsvalues for A [burdenNumericalAnalysis2016]_.

NumericalMethods.MultiVariableIterations.successive_relaxation — Function

successive_relaxation(mvi[, omega=0.])

Solve $\vec{x} = \mathbf{A}^{-1}\vec{b}$ via the Successive Relaxation Method. Method is Successive Over-Relaxation (SOR) if omega > 1, Successive Under-Relaxation (SUR) if omega < 1, and reduces to Gauss-Seidel if omega = 1.

Notes

SOR and SUR accelerate or deccelerate convergence of gauss_seidel(), respectively, by decreasing or increasing the spectral radius of mvi.A. The core algorithm by which method marches through iterations:

\[ \vec{x}^{(k)} = \bigl( (\mathbf{D} - ω\mathbf{L})^{-1} * ((1 - ω)*\mathbf{D} + ω\mathbf{U}) \bigr) ⋅ \vec{x}^{(k - 1)} + ω( (\mathbf{D} - ω\mathbf{L})^{-1} ) ⋅ \vec{b}\]

If left unspecified, and if possible, an optimum relaxation parameter, ω will be calculated by find_omega().

NumericalMethods.MultiVariableIterations.symmetry — Method

symmetry(A)

Determines whether matrix, A is symmetric.

NumericalMethods.MultiVariableIterations.tridiagonality — Method

tridiagonality(A)

Determine whether matrix, A is tridiagonal.

NumericalMethods.newton_raphson — Method

newton_raphson(mvi::MultiVariableIteration, variables::Tuple{Vararg{Num}}[; jacobian=nothing])

Solve non-linear systems of equations, $\vec{x} = \mathbf{A}^{-1}\vec{b}$ via the Newton-Raphson Method.

Here, mvi.A should be a vector of functions wherein each variable is represented. Method will go faster if jacobian is pre-defined. Otherwise, the Jacobian matrix of mvi.A will be internally constructed.

Examples

f1(x1, x2, x3)  = 3x1 - cos(x2*x3) - 0.5
f2(x1, x2, x3)  = x1^2 - 81(x2 + 0.1)^2 + sin(x3) + 1.06
f3(x1, x2, x3)  = exp(-x1*x2) + 20x3 + 3\(10π - 3)
A               = [f1, f2, f3]
b               = zeros(length(A))
x0              = [0.1, 0.1, -0.1]
tol             = 1e-9
mvi             = MultiVariableIteration(A, x0, b, 5, tol)
using Symbolics
@variables x1, x2, x3
newton_raphson(mvi, (x1, x2, x3))

NumericalMethods.solve — Method

solve(mvi::MultiVariableIteration[; method=:jacobi, omega=0., variables, jacobian])

Solve $\vec{x} = \mathbf{A}^{-1}\vec{b}$ according to method ∈ {:jacobi (default), :gauss_seidel, :successive_relaxation, :newton_raphson}.

Each method has an equivalent convenience function. E.g. solve(mvi; method=:jacobi) ≡ jacobi(mvi).

Index

NumericalMethods.MultiVariableIterations.MultiVariableIteration

NumericalMethods.MultiVariableIterations.condition_number

NumericalMethods.MultiVariableIterations.diagonality

NumericalMethods.MultiVariableIterations.find_omega

NumericalMethods.MultiVariableIterations.gauss_seidel

NumericalMethods.MultiVariableIterations.jacobi

NumericalMethods.MultiVariableIterations.positive_definite

NumericalMethods.MultiVariableIterations.spectral_radius

NumericalMethods.MultiVariableIterations.successive_relaxation

NumericalMethods.MultiVariableIterations.symmetry

NumericalMethods.MultiVariableIterations.tridiagonality