Numerical Differentiation

Functions

NumericalMethods.Derivatives.endpoint — Method

endpoint(x, f, h, point[; method=:three])

Find the derivative of a bookend point at either :begin or :end of dataset. Acceptable values for method include {:three, :five}.

NumericalMethods.Derivatives.midpoint — Method

midpoint(x, f, h, point[; method=:three])

Find the derivative of some point within a dataset. Acceptable values for method include {:three, :five, :2nd}.

NumericalMethods.Derivatives.n1derivative — Method

n1derivative(x, f, j[; n=nothing])

The general (n + 1)-point formula to approximate f' at point j.

Notes

If n = nothing, then entire dataset used to construct n'th Lagrange coefficient.

Index

NumericalMethods.Derivatives.endpoint

NumericalMethods.Derivatives.midpoint

NumericalMethods.Derivatives.n1derivative