NumericalMethods

Documentation for NumericalMethods.

NumericalMethods

Functions

Index

Functions

NumericalMethods.InitialValueProblem — Type

InitialValueProblem(f, a, b, h, α, N)

Structure of the boundary conditions to Initial-Value Problem (IVP) differential equation.

Notes

Make sure the independent variable (e.g. time) is the first argument of f!

NumericalMethods.MultiVariableIteration — Type

MultiVariableIteration(A, x, b, N, tol)

Given the system of equations, $\mathbf{A}\vec{x} = \vec{b}$, attempt to solve $\vec{x} = \mathbf{A}^{-1}\vec{b}$ in so many iterations, N within tolerance, tol. Ideal for large, sparse systems.

NumericalMethods.SingleVariableIteration — Type

SingleVariableIteration(f, a, b, n, tol)

Given f(p) such that p ∈ [a, b], find the root of a single-variable equation in so many iterations, n within tolerance, tol.

NumericalMethods.bezier — Method

bezier(x, y, xguides, yguides)

An application of Hermitic polynomials to draw Bezier curves between points.

Notes

Each argument should be a one-to-one mapping of points, (xᵢ, yᵢ) and (xᵢ₊₁, yᵢ₊₁) and their respective guide points, (xᵢ⁺, yᵢ⁺) and (xᵢ₊₁⁻, yᵢ₊₁⁻).

NumericalMethods.bisection — Method

bisection(svi::SingleVariableIteration)

Root-finding method: f(x) = 0. Search for solution by halving the bounds such that a and b initially yield opposite signs in function.

Notes

Relying on the Intermediate Value Theorem (IVT), this is a bracketed, root-finding method. This method is rather slow to converge but will always converge to a solution; therefore, is a good starter method.

NumericalMethods.clamped — Method

clamped(x, f, fp)

The bookend polynomials will have the same slope entering and exiting the interval as the derivative at the respective endpoint.

NumericalMethods.condition_number — Method

condition_number(A)

Find the condition number of matrix, A.

Notes

Definition [burdenNumericalAnalysis2016]_: The condition number of the non-singular matrix, $\mathbf{A}$ relative to a norm, ||⋅|| is

\[ K(\mathbf{A}) = ||\mathbf{A}|| ⋅ ||\mathbf{A}^{-1}||\]

A matrix is well-conditioned if $K(\mathbf{A})$ is close to 1 and is ill-conditioned if significantly greater than 1.

NumericalMethods.diagonality — Method

diagonality(A)

Determines whether matrix, A is strictly, diagonally dominant.

NumericalMethods.endpoint — Method

endpoint(x, f, h, point[; method=:three])

Find the derivative of a bookend point at either :begin or :end of dataset. Acceptable values for method include {:three, :five}.

NumericalMethods.false_position — Method

false_position(svi::SingleVariableIteration, p0, p1)

Attempt root-finding method with initial guesses such that p0 and p1 in [a, b] yield opposite signs in function.

Use function with lowest slope!

Notes

Similar to, but slower to converge than, the secant_method() by including a test to ensure solution is root-bracketed.

See fixed_point() for theorem.

NumericalMethods.find_omega — Function

find_omega(mvi[, omega=0.])

Find the optimum relaxation parameter, omega for mvi.A and mvi.x, if possible.

Notes

If 0 < omega < 2, then method will converge regardless of choice for mvi.x. If matrix, mvi.A is tridiagonal, then the spectral radius of Gauss-Seidel's T-matrix, $\mathbf{T}\_{g}$ is used to calculate $ω := 2 / (1 + \sqrt{(1 - `spectral_radius`(\mathbf{T}\_{g})})$ [burdenNumericalAnalysis2016]_.

NumericalMethods.fixed_point — Method

fixed_point(svi::SingleVariableIteration, p0)

Attempt root-finding method with initial guess, p0 in [a, b] by solving the equation g(p) = p via f(p) - p = 0.

Use function with lowest slope!

Not root-bracketed.

Notes

Theorem:

Existence of a fixed-point: If g ∈ C[a,b] and g(x) ∈ C[a, b] for all x ∈ [a, b], then function, g has a fixed point, p ∈ [a, b].

Uniqueness of a fixed point: If g'(x) exists on [a, b] and a positive constant, k < 1 exist with {|g'(x)| ≤ k | x ∈ (a, b)}, then there is exactly one fixed-point, p ∈ [a, b].

Converges by $\mathcal{O}(\text{linear})$ if g'(p) ≠ 0, and $\mathcal{O}(\text{quadratic})$ if g'(p) = 0 and g''(p) < M, where M = g''(ξ) that is the error function.

NumericalMethods.gauss_seidel — Method

gauss_seidel(mvi)

Solve $\vec{x} = \mathbf{A}^{-1}\vec{b}$ via the Gauss-Seidel Method.

Notes

This method improves on jacobi() by using the most recently calculated entries in the approximation vector, x at the end of each iteration. The core algorithm by which method marches through iterations:

\[ \vec{x}^{(k)} = \bigl( (\mathbf{D} - \mathbf{L})^{-1} * \mathbf{U} \bigr) ⋅ \vec{x}^{(k - 1)} + \bigl( (\mathbf{D} - \mathbf{L})^{-1} \bigr) ⋅ \vec{b}\]

NumericalMethods.integrate — Method

integrate(f, x      [; rule=:trapezoidal, tol=10^-3])
integrate(f, a, b, h[; rule=:trapezoidal, tol=10^-3])
integrate(f, a, b, n[; rule=:trapezoidal, tol=10^-3])

Find the definite integral by some numerical quadrature.

Notes

f may be a function or vector. The domain may be defined with a vector, x or on the interval [a, b] either by number of sub-intervals, n or step-size, h. rule accepts {:trapezoidal (default), :midpoint, :simpson13, :simpson38, :simpsonN}. Dataset may contain unevenly spaces points.

References

https://en.wikipedia.org/wiki/Simpson%27s_rule

NumericalMethods.jacobi — Method

jacobi(mvi)

Solve $\vec{x} = \mathbf{A}^{-1}\vec{b}$ via the Jacobi Method to find $\vec{x}$.

Notes

The core algorithm by which method marches through iterations:

\[ \vec{x}^{(k)} = \bigl( \mathbf{D}^{-1} * (\mathbf{L} + \mathbf{U}) \bigr) ⋅ \vec{x}^{(k - 1)} + ( \mathbf{D}^{-1} ) ⋅ \vec{b}\]

NumericalMethods.lagrange — Method

lagrange(x, f[; n=nothing])

Given a domain, x and range, f, construct the nth Lagrangian polynomial.

Notes

If n=nothing, then method will utilize entire dataset. Polynomials will quickly oscillate for larger datasets.

NumericalMethods.linearleastsquares — Method

linearleastsquares(x, f, n::Int64)

Construct a polynomial of degree, n while minimizing the least squares error.

Notes

Least squares error := $E = \sum_{i=1}^{m}[y_{i} - P_{n}(x_{i})]^{2}$

Constructed polynomial of the form: $P(x) = a_{n}x^{n} + a_{n - 1}x^{n - 1} + \dots + a_{1}x + a_{0}$

NumericalMethods.linearleastsquares — Method

linearleastsquares(x, f, type::Symbol)

Given a domain and range, yield the coefficients for an equation and the equation of the form $y = ax^{b}$.

NumericalMethods.maximum_iterations — Function

maximum_iterations(obj, method[, p0=0])

Find greatest integer for maximum iterations within tolerance.

Notes

Acceptable values for method ∈ {:bisection, :fixed_point, :newton_raphson, :secant_method, :false_position}. Initial guess, p0 for function solution is not needed for :bisection method.

NumericalMethods.maximum_slope — Method

maximum_slope(svi::SingleVariableIteration)

Find the greatest value for first derivative of function.

NumericalMethods.midpoint — Method

midpoint(x, f, h, point[; method=:three])

Find the derivative of some point within a dataset. Acceptable values for method include {:three, :five, :2nd}.

NumericalMethods.n1derivative — Method

n1derivative(x, f, j[; n=nothing])

The general (n + 1)-point formula to approximate f' at point j.

Notes

If n = nothing, then entire dataset used to construct n'th Lagrange coefficient.

NumericalMethods.natural — Method

natural(x, f)

The bookend polynomials do not assume the slope entering and exiting the interval as the derivative at the respective endpoint.

NumericalMethods.newton_raphson — Method

newton_raphson(mvi::MultiVariableIteration, variables::Tuple{Vararg{Num}})

Solve non-linear systems of equations, $\vec{x} = \mathbf{A}^{-1}\vec{b}$ via the Newton-Raphson Method.

Here, mvi.A should be a vector of functions wherein each variable is represented.

Examples

f1(x1, x2, x3)  = 3x1 - cos(x2*x3) - 0.5
f2(x1, x2, x3)  = x1^2 - 81(x2 + 0.1)^2 + sin(x3) + 1.06
f3(x1, x2, x3)  = exp(-x1*x2) + 20x3 + 3\(10π - 3)
A               = [f1, f2, f3]
b               = zeros(length(A))
x0              = [0.1, 0.1, -0.1]
tol             = 1e-9
mvi             = MultiVariableIteration(A, x0, b, 5, tol)
using Symbolics
@variables x1, x2, x3
newton_raphson(mvi, (x1, x2, x3))

NumericalMethods.newton_raphson — Method

newton_raphson(svi::SingleVariableIteration, p0[; df=nothing])

Attempt root-finding method with initial guess, p0 in [a, b] by solving the equation g(p) = p via f(p) - p = 0. df will be the first derivative of function if not given.

Use function with lowest slope!

df(x) ≠ 0

Notes

Quickest convergence rate, but not root-bracketed and has trouble with symmetric functions! Initial guess, p0 must be close to real solution; else, will converge to different root or oscillate (if symmetric). This method can be viewed as fixed-point iteration.

Technique based on first Taylor polynomial expansion of f about $p_{0}$ (that is p0) and evaluated at x = p. |p - $p_{0}$| is assumed small; therefore, 2ⁿᵈ-order Taylor term, the error, is small.

See fixed_point() for theorem.

NumericalMethods.newtondifference — Method

newtondifference(x, f, α[; dir::Symbol=:auto])

Given a domain, x and range, f, construct some polynomial by Newton's Divided Difference centered around α. :forward or :backward construction.

Notes

Direction will be chosen if not specified. Polynomials best made with even spacing in x; although, this is not completely necessary.

NumericalMethods.positive_definite — Method

positive_definite(A)

Determines whether matrix, A is positive definite.

NumericalMethods.secant_method — Method

secant_method(svi::SingleVariableIteration, p0, p1)

Attempt root-finding method with initial guesses such that p0 and p1 in [a, b] yield opposite signs in function.

Use function with lowest slope!

Not root-bracketed.

Notes

Method is less computationally expensive than newton_raphson() but may converge at slower rate by circumventing need to calculate derivative.

See fixed_point() for theorem.

NumericalMethods.solve — Method

solve(ivp::InitialValueProblem[; method=:forward_euler, tol=10^-3])

Solve ivp according to method ∈ {:forward_euler (default), :backward_euler, :improved_euler, :modified_euler, :runge_kutta}.

Notes

Each method has an equivalent convenience function. E.g. solve(ivp; method=:runge_kutta) ≡ runge_kutta(ivp).

NumericalMethods.solve — Method

solve(mvi::MultiVariableIteration[; method=:jacobi, omega=0., variables])

Solve $\vec{x} = \mathbf{A}^{-1}\vec{b}$ according to method ∈ {:jacobi (default), :gauss_seidel, :successive_relaxation, :newton_raphson}.

Each method has an equivalent convenience function. E.g. solve(mvi; method=:jacobi) ≡ jacobi(ivp).

NumericalMethods.solve — Method

solve(svi::SingleVariableIteration[; method=:bisection, p0, p1, df])

Attempt to find where f(p) = 0 according to method ∈ {:bisection (default), :fixed_point, :newton_raphson, :secant_method, :false_position}. Each method also has a convenience function.

Notes

Convergence Rates: :newton_raphson > :secant_method > :false_position > :fixed_point > :bisection

:bisection is default because will always converge to a solution but is not the quickest.

NumericalMethods.spectral_radius — Method

spectral_radius(A)

Finds the spectral radius of matrix, A.

Notes

$ρ(\mathbf{A}) = \max|λ|$, where λ is the set of eigvalsvalues for A [burdenNumericalAnalysis2016]_.

NumericalMethods.successive_relaxation — Function

successive_relaxation(mvi[, omega=0.])

Solve $\vec{x} = \mathbf{A}^{-1}\vec{b}$ via the Successive Relaxation Method. Method is Successive Over-Relaxation (SOR) if omega > 1, Successive Under-Relaxation (SUR) if omega < 1, and reduces to Gauss-Seidel if omega = 1.

Notes

SOR and SUR accelerate or deccelerate convergence of gauss_seidel(), respectively, by decreasing or increasing the spectral radius of mvi.A. The core algorithm by which method marches through iterations:

\[ \vec{x}^{(k)} = \bigl( (\mathbf{D} - ω\mathbf{L})^{-1} * ((1 - ω)*\mathbf{D} + ω\mathbf{U}) \bigr) ⋅ \vec{x}^{(k - 1)} + ω( (\mathbf{D} - ω\mathbf{L})^{-1} ) ⋅ \vec{b}\]

If left unspecified, and if possible, an optimum relaxation parameter, ω will be calculated by find_omega().

NumericalMethods.symmetry — Method

symmetry(A)

Determines whether matrix, A is symmetric.

NumericalMethods.tridiagonality — Method

tridiagonality(A)

Determine whether matrix, A is tridiagonal.

Index

NumericalMethods.InitialValueProblem

NumericalMethods.MultiVariableIteration

NumericalMethods.SingleVariableIteration

NumericalMethods.bezier

NumericalMethods.bisection

NumericalMethods.clamped

NumericalMethods.condition_number

NumericalMethods.diagonality

NumericalMethods.endpoint

NumericalMethods.false_position

NumericalMethods.find_omega

NumericalMethods.fixed_point

NumericalMethods.gauss_seidel

NumericalMethods.integrate

NumericalMethods.jacobi

NumericalMethods.lagrange

NumericalMethods.linearleastsquares

NumericalMethods.linearleastsquares

NumericalMethods.maximum_iterations

NumericalMethods.maximum_slope

NumericalMethods.midpoint

NumericalMethods.n1derivative

NumericalMethods.natural

NumericalMethods.newton_raphson

NumericalMethods.newton_raphson

NumericalMethods.newtondifference

NumericalMethods.positive_definite

NumericalMethods.secant_method

NumericalMethods.solve

NumericalMethods.solve

NumericalMethods.solve

NumericalMethods.spectral_radius

NumericalMethods.successive_relaxation

NumericalMethods.symmetry

NumericalMethods.tridiagonality