NumericalMethods

Documentation for NumericalMethods.

NumericalMethods

Functions

Index

Functions

NumericalMethods.ODE — Type

ODE(f, a, b, h, α, N)

Structure of the boundary conditions to differential equation where f is the time derivative of the function to approximate.

Notes

Make sure the independent variable is the first argument of f!

NumericalMethods.clamped — Method

clamped(domain, f, function_derivative)

The bookend polynomials will have the same slope entering and exiting the interval as the derivative at the respective endpoint.

NumericalMethods.endpoint — Method

endpoint(x, y, h, point[, point_type="three"])

Find the derivative of a bookend point at either :begin or :end of dataset. Acceptable values for point_type include {"three", "five"}.

NumericalMethods.integrate — Method

integrate(f[; rule=:trapezoidal, tol=10^-3])

Notes

Find the definite integral by some composite numeric quadrature. f may be a function or range. The domain may be defined with a vector, x or on the interval [a, b] either by number of sub-intervals, n or step-size, h. rule accepts {:trapezoidal (default), :midpoint, :simpson13, :simpson38, :simpsonN}. Dataset may contain unevenly spaces points.

References

https://en.wikipedia.org/wiki/Simpson%27s_rule

NumericalMethods.ivp — Function

ivp(obj::ODE[, tol=10^-3; method=:forward_euler])

Solve obj according to method ∈ {:forward_euler (default), :backward_euler, :improved_euler, :modified_euler, :runge_kutta}.

Notes

Each method has an equivalent convenience function. E.g. ivp(obj; method=:runge_kutta) ≡ runge_kutta(obj).

NumericalMethods.lagrange — Method

lagrange()

Given a domain and range, construct a Lagrangian polynomial. Polynomial will quickly begin to oscillate for larger datasets.

NumericalMethods.linearleastsquares — Method

linearleastsquares(domain, f, degree::Integer)

Construct a polynomial of some degree while minimizing the least squares error.

Notes

Least squares error := $E = \sum_{i=1}^{m}[y_{i} - P_{n}(x_{i})]^{2}$

Constructed polynomial of the form: $P(x) = a_{n}x^{n} + a_{n - 1}x^{n - 1} + \dots + a_{1}x + a_{0}$

NumericalMethods.linearleastsquares — Method

linearleastsquares(domain, f, type::Symbol)

Given a domain and range, yield the coefficients for an equation and the equation of the form $y = ax^{b}$.

NumericalMethods.midpoint — Method

midpoint(x, y, h, point[, point_type="three"])

Find the derivative of some point within a dataset. Acceptable values for point_type include {"three", "five", "2nd_derivative"}.

NumericalMethods.natural — Method

natural(domain, f, function_derivative)

The bookend polynomials do not assume the slope entering and exiting the interval as the derivative at the respective endpoint.

NumericalMethods.newtondifference — Method

newtondifference(x, f, α[; var::Symbol=:x, dir::Symbol=:auto])

Given a domain and range, construct some polynomial by Newton's Divided Difference. 'forward' or 'backward' construction. Will be chosen automatically if not specified.

Notes

Direction will be chosen if not specified. Polynomials best made with even spacing in domain; although, this is not completely necessary.

Index

NumericalMethods.ODE

NumericalMethods.clamped

NumericalMethods.endpoint

NumericalMethods.integrate

NumericalMethods.ivp

NumericalMethods.lagrange

NumericalMethods.linearleastsquares

NumericalMethods.linearleastsquares

NumericalMethods.midpoint

NumericalMethods.natural

NumericalMethods.newtondifference